refactor: VoteFormModal 로직 분리 및 유틸 함수 정리
