+++ src/utils/dateFormatter.ts
@@ -2,3 +2,8 @@ export const formatTime = (iso: string) => {
   const d = new Date(iso);
   return `${d.getMonth() + 1}월 ${d.getDate()}일 ${d.getHours()}시`;
 };
+
+export const toDatetimeLocalFormat = (date: Date): string => {
+  const pad = (n: number) => n.toString().padStart(2, '0');
+  return `${date.getFullYear()}-${pad(date.getMonth() + 1)}-${pad(date.getDate())}T${pad(date.getHours())}:${pad(date.getMinutes())}`;
+};
